+++ annotations/src/main/java/arez/annotations/ComponentNameRef.java
@@ -15,7 +15,7 @@ import java.lang.annotation.Target;
  * annotation should be on a method that accepts no parameters and returns
  * a String.</p>
  *
- * <p>The method that is annotated with @ComponentName must comply with the additional constraints:</p>
+ * <p>The method that is annotated with @ComponentNameRef must comply with the additional constraints:</p>
  * <ul>
  * <li>Must not be annotated with any other arez annotation</li>
  * <li>Must have 0 parameters</li>
@@ -27,10 +27,10 @@ import java.lang.annotation.Target;
  * <li>Must not throw exceptions</li>
  * </ul>
  * @see ComponentId
- * @see ComponentTypeName
+ * @see ComponentTypeNameRef
  */
 @Documented
 @Target( ElementType.METHOD )
-public @interface ComponentName
+public @interface ComponentNameRef
 {
 }