+++ annotations/src/main/java/arez/annotations/ComponentTypeNameRef.java
@@ -9,13 +9,13 @@ import org.realityforge.anodoc.Unsupported;
  * Annotate the method that will be overridden to return the value of ArezComponent.name().
  * This is only useful when a class potentially has multiple {@link ArezComponent} annotated subclasses.
  * This method is used when manually setting up debug context and you want it to align with the concrete
- * implementations. However often {@link ComponentName} is a better solution. If not specified Arez will
+ * implementations. However often {@link ComponentNameRef} is a better solution. If not specified Arez will
  * generate a private method if needed.
  *
  * <p>This annotation should appear at most once on a component. The annotation should be on a method that
  * can be overridden, accepts no parameters and returns a String.</p>
  *
- * <p>The method that is annotated with @ComponentTypeName must comply with the additional constraints:</p>
+ * <p>The method that is annotated with @ComponentTypeNameRef must comply with the additional constraints:</p>
  * <ul>
  * <li>Must not be annotated with any other arez annotation</li>
  * <li>Must have 0 parameters</li>
@@ -30,6 +30,6 @@ import org.realityforge.anodoc.Unsupported;
 @Documented
 @Target( ElementType.METHOD )
 @Unsupported( "The utility of this annotation is in question and may be removed in a future version of the library" )
-public @interface ComponentTypeName
+public @interface ComponentTypeNameRef
 {
 }