+++ processor/src/test/resources/bad_input/com/example/component_name/ComponentNameDuplicateModel.java
@@ -2,7 +2,7 @@ package com.example.component_name;
 
 import arez.annotations.Action;
 import arez.annotations.ArezComponent;
-import arez.annotations.ComponentName;
+import arez.annotations.ComponentNameRef;
 
 @ArezComponent
 public class ComponentNameDuplicateModel
@@ -12,13 +12,13 @@ public class ComponentNameDuplicateModel
   {
   }
 
-  @ComponentName
+  @ComponentNameRef
   String getTypeName()
   {
     return null;
   }
 
-  @ComponentName
+  @ComponentNameRef
   String getTypeName2()
   {
     return null;