+++ processor/src/test/resources/bad_input/com/example/component_name/ComponentNameStaticModel.java
@@ -2,7 +2,7 @@ package com.example.component_name;
 
 import arez.annotations.Action;
 import arez.annotations.ArezComponent;
-import arez.annotations.ComponentName;
+import arez.annotations.ComponentNameRef;
 
 @ArezComponent
 public class ComponentNameStaticModel
@@ -12,7 +12,7 @@ public class ComponentNameStaticModel
   {
   }
 
-  @ComponentName
+  @ComponentNameRef
   static String getTypeName()
   {
     return null;