+++ processor/src/test/resources/bad_input/com/example/component_name/ComponentNameThrowsExceptionModel.java
@@ -2,7 +2,7 @@ package com.example.component_name;
 
 import arez.annotations.Action;
 import arez.annotations.ArezComponent;
-import arez.annotations.ComponentName;
+import arez.annotations.ComponentNameRef;
 import java.io.IOException;
 
 @ArezComponent
@@ -13,7 +13,7 @@ public class ComponentNameThrowsExceptionModel
   {
   }
 
-  @ComponentName
+  @ComponentNameRef
   String getTypeName()
     throws IOException
   {