+++ processor/src/test/resources/bad_input/com/example/component_type_name/ComponentTypeNameDuplicateModel.java
@@ -2,7 +2,7 @@ package com.example.component_type_name;
 
 import arez.annotations.Action;
 import arez.annotations.ArezComponent;
-import arez.annotations.ComponentTypeName;
+import arez.annotations.ComponentTypeNameRef;
 
 @ArezComponent
 public class ComponentTypeNameDuplicateModel
@@ -12,13 +12,13 @@ public class ComponentTypeNameDuplicateModel
   {
   }
 
-  @ComponentTypeName
+  @ComponentTypeNameRef
   String getTypeName()
   {
     return null;
   }
 
-  @ComponentTypeName
+  @ComponentTypeNameRef
   String getTypeName2()
   {
     return null;