+++ processor/src/test/resources/bad_input/com/example/component_type_name/ComponentTypeNamePrivateModel.java
@@ -2,7 +2,7 @@ package com.example.component_type_name;
 
 import arez.annotations.Action;
 import arez.annotations.ArezComponent;
-import arez.annotations.ComponentTypeName;
+import arez.annotations.ComponentTypeNameRef;
 
 @ArezComponent
 public class ComponentTypeNamePrivateModel
@@ -12,7 +12,7 @@ public class ComponentTypeNamePrivateModel
   {
   }
 
-  @ComponentTypeName
+  @ComponentTypeNameRef
   private String getTypeName()
   {
     return null;